requirements.txt
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.2.0
pytesseract==0.3.5
requests==2.24.0
soupsieve==2.0.1
urllib3==1.25.10
Werkzeug==1.0.1
